Version 1.9.1
make mask in lexpermute_seven into a bit sequence.  Only about a 7% speed
increase on average, but up to 15% in some places.  Generally try to
minimize memory footprint in the bulk of the subroutine.

Lots of debug code in here because this took awhile to figure out
correctly.  Will remove it in next revision.